fallos de seguridad

+++ src/controllers/tasks/deleteTask.controller.js
@@ -1,8 +1,12 @@
+import verifyOwner from "../../middlewares/verifyOwner.js";
 import { deleteTaskModel } from "../../models/tasks/index.js";
 import deleteTaskSchema from "../../validations/deleteTaskSchema.js";
+import verifyTask from "../../validations/verifyTask.js";
 
 export default async function deleteTaskController(req, res, next) {
   try {
+    const currentUser = req.currentUser.id;
+
     const { error, value } = deleteTaskSchema.validate({
       task_id: req.params.idTask,
     });
@@ -15,7 +19,14 @@ export default async function deleteTaskController(req, res, next) {
     }
 
     const { task_id } = value;
-
+    const [[task]] = await verifyTask(task_id);
+    if (!task) {
+      return res.status(400).json({
+        ok: false,
+        message: "task no encontrado",
+      });
+    }
+    verifyOwner(task, currentUser);
     const { message } = await deleteTaskModel(task_id);
     return res.status(200).json({
       ok: true,
